gmscompat: refactoring for Android 13
android.content.pm.parsing package moves to com.android.server package and is no longer accessible
from com.android.internal.
