+++ services/core/java/com/android/server/pm/pkg/parsing/ParsingPackageUtils.java
@@ -88,11 +88,10 @@ import android.util.TypedValue;
 import android.util.apk.ApkSignatureVerifier;
 
 import com.android.internal.R;
-import com.android.internal.gmscompat.PlayStoreHooks;
-import com.android.internal.gmscompat.client.GmsClientHooks;
 import com.android.internal.os.ClassLoaderFactory;
 import com.android.internal.util.ArrayUtils;
 import com.android.internal.util.XmlUtils;
+import com.android.server.ext.GmsSysServerHooks;
 import com.android.server.pm.SharedUidMigration;
 import com.android.server.pm.permission.CompatibilityPermissionInfo;
 import com.android.server.pm.pkg.component.ComponentMutateUtils;
@@ -2197,8 +2196,8 @@ public class ParsingPackageUtils {
             pkg.addActivity(a.getResult());
         }
 
-        GmsClientHooks.maybeAddServiceDuringParsing(pkg);
-        PlayStoreHooks.maybeAddUsesPermission(pkg);
+        GmsSysServerHooks.maybeAddServiceDuringParsing(pkg);
+        GmsSysServerHooks.maybeAddUsesPermission(pkg);
 
         if (hasActivityOrder) {
             pkg.sortActivities();
